Make MadgraphTest header only.
This allows for pulling the compilation of tests down into each implementation
directory, directly using the floating-point type that madgraph uses.
This also allowed for removing some test instantiation logic.

Adapt epochX:
Since there is no more object file in the test/ directory, the
Makefiles of the cudacpp codes can be simplified a bit:
For tests, they now need to ensure the presence of googletest and add
an include statement for MadgraphTest.h. The compilation now happens
entirely by including a header.
