+++ test/include/MadgraphTest.h
@@ -2,19 +2,78 @@
 #ifndef MADGRAPHTEST_H_
 #define MADGRAPHTEST_H_ 1
 
+#include <mgOnGpuConfig.h>
+
 #include <array>
+#include <cmath>
+#include <fstream>
 #include <iomanip>
+#include <memory>
+#include <sstream>
 #include <string>
 #include <vector>
 
 #include <gtest/gtest.h>
 
+namespace {
+
 struct ReferenceData {
-  std::vector< std::vector<std::array<double, 4>> > momenta;
+  std::vector< std::vector<std::array<double, mgOnGpu::np4>> > momenta;
   std::vector<double> MEs;
 };
 
-std::map<unsigned int, ReferenceData> readReferenceData(const std::string& refFileName);
+/// Read batches of reference data from a file and store them in a map.
+std::map<unsigned int, ReferenceData> readReferenceData(const std::string& refFileName)
+{
+  std::ifstream referenceFile(refFileName.c_str());
+  EXPECT_TRUE(referenceFile.is_open()) << refFileName;
+  std::map<unsigned int, ReferenceData> referenceData;
+  unsigned int evtNo;
+  unsigned int batchNo;
+
+  for (std::string line; std::getline(referenceFile, line); )
+  {
+    std::stringstream lineStr(line);
+    if (line.empty() || line[0] == '#')
+    {
+      continue;
+    }
+    else if (line.find("Event") != std::string::npos)
+    {
+      std::string dummy;
+      lineStr >> dummy >> evtNo >> dummy >> batchNo;
+    }
+    else if (line.find("ME") != std::string::npos)
+    {
+      if (evtNo <= referenceData[batchNo].MEs.size())
+        referenceData[batchNo].MEs.resize(evtNo + 1);
+
+      std::string dummy;
+      lineStr >> dummy >> referenceData[batchNo].MEs[evtNo];
+    }
+    else
+    {
+      unsigned int particleIndex;
+      lineStr >> particleIndex;
+
+      if (evtNo <= referenceData[batchNo].momenta.size())
+        referenceData[batchNo].momenta.resize(evtNo + 1);
+      if (particleIndex <= referenceData[batchNo].momenta[evtNo].size())
+        referenceData[batchNo].momenta[evtNo].resize(particleIndex + 1);
+
+      auto& fourVec = referenceData[batchNo].momenta[evtNo][particleIndex];
+      for (unsigned int i=0; i < fourVec.size(); ++i) {
+        EXPECT_TRUE(lineStr.good());
+        lineStr >> fourVec[i];
+      }
+      EXPECT_TRUE(lineStr.eof());
+    }
+  }
+  return referenceData;
+}
+
+}
+
 
 /**
  * Test driver providing a common interface for testing different implementations.
@@ -24,60 +83,73 @@ std::map<unsigned int, ReferenceData> readReferenceData(const std::string& refFi
  *
  * Usage:
  * ```
- * class TestImplementation : public BaseTest {
- *   <implement functions>
+ * class TestImplementation : public TestDriverBase {
+ *   <override all pure-virtual functions with Madgraph workflow>
  * }
  *
- * TEST_F(TestImplementation, <testName>) {
- *   <test code>
+ * class TestImplementation2 : public TestDriverBase {
+ *   <override all pure-virtual functions with a different Madgraph workflow>
  * }
+ *
+ * INSTANTIATE_TEST_SUITE_P( TestName,
+ *                           MadgraphTest,
+ *                           testing::Values( new TestImplementation, new TestImplementation2, ... ) );
+ *```
+ *
+ * For adapting the test workflow, see the .cc and adapt
+ *   TEST_P(MadgraphTest, CompareMomentaAndME)
+ *
+ * To add a test that should be runnable with all test implementations that derive from TestDriverBase, add a new
+ *   TEST_P(MadgraphTest, <TestName>) {
+ *     <test code>
+ *   }
  */
-template<typename Fptype>
 class TestDriverBase {
-private:
   std::string m_refFileName;
+
 public:
-  unsigned int nparticle = 4; // NB This is a (non-static, non-const) instance variable which gets modified in runTest.cc
-  static constexpr unsigned int np4 = 4;
+  const unsigned int nparticle;
   static constexpr unsigned int niter = 2;
   static constexpr unsigned int gpublocks = 2;
   static constexpr unsigned int gputhreads = 128;
   static constexpr unsigned int nevt = gpublocks * gputhreads;
 
-  TestDriverBase( const std::string& refFileName = "" ) // default value "" is for backward compatibility
-    : m_refFileName( refFileName != "" ? refFileName : std::string( "../../../../../test/ref/dump_CPUTest.Sigma_sm_epem_mupmum.txt" ) ){}
+  TestDriverBase( unsigned int npart = mgOnGpu::npar, const std::string& refFileName = "" ) // default value "" is for backward compatibility
+    : m_refFileName( refFileName != "" ? refFileName : std::string( "../../../../../test/ref/dump_CPUTest.Sigma_sm_epem_mupmum.txt" ) ),
+      nparticle{ npart } {}
   virtual ~TestDriverBase() { }
-  const std::string& getRefFileName(){ return m_refFileName; }
+
+  const std::string & getRefFileName() const { return m_refFileName; }
 
   // ------------------------------------------------
   // Interface for retrieving info from madgraph
   // ------------------------------------------------
-  virtual Fptype getMomentum(std::size_t evtNo, unsigned int particleNo, unsigned int component) const = 0;
-  virtual Fptype getMatrixElement(std::size_t evtNo) const = 0;
+  virtual fptype getMomentum(std::size_t evtNo, unsigned int particleNo, unsigned int component) const = 0;
+  virtual fptype getMatrixElement(std::size_t evtNo) const = 0;
 
   // ------------------------------------------------
   // Interface for steering madgraph run
   // ------------------------------------------------
   virtual void prepareRandomNumbers(unsigned int iiter) = 0;
-  virtual void prepareMomenta(Fptype energy) = 0;
+  virtual void prepareMomenta(fptype energy) = 0;
   virtual void runSigmaKin(std::size_t iiter) = 0;
 
-  // Print the requested event into the stream. If the reference data has enough events, it will be printed as well.
+  /// Print the requested event into the stream. If the reference data has enough events, it will be printed as well.
   void dumpParticles(std::ostream& stream, std::size_t ievt, unsigned int numParticles,
-                     unsigned int nDigit, const ReferenceData& referenceData)
+                     unsigned int nDigit, const ReferenceData& referenceData) const
   {
     const auto width = nDigit + 8;
     for (unsigned int ipar = 0; ipar < numParticles; ipar++)
     {
       // NB: 'setw' affects only the next field (of any type)
       stream << std::scientific // fixed format: affects all floats (default nDigit: 6)
-             << std::setprecision(nDigit)
-             << std::setw(4) << ipar
-             << std::setw(width) << getMomentum(ievt, ipar, 0)
-             << std::setw(width) << getMomentum(ievt, ipar, 1)
-             << std::setw(width) << getMomentum(ievt, ipar, 2)
-             << std::setw(width) << getMomentum(ievt, ipar, 3)
-             << "\n";
+          << std::setprecision(nDigit)
+          << std::setw(4) << ipar
+          << std::setw(width) << getMomentum(ievt, ipar, 0)
+          << std::setw(width) << getMomentum(ievt, ipar, 1)
+          << std::setw(width) << getMomentum(ievt, ipar, 2)
+          << std::setw(width) << getMomentum(ievt, ipar, 3)
+          << "\n";
       if (ievt < referenceData.momenta.size()) {
         stream << "ref" << ipar;
         stream << std::setw(width) << referenceData.momenta[ievt][ipar][0]
@@ -88,22 +160,143 @@ public:
       }
       stream << std::flush << std::defaultfloat; // default format: affects all floats
     }
-  };
+  }
 };
 
-// Test class that's using the driver to run the test(s) below.
-template<typename Fptype>
-class MadgraphTestFptype : public testing::TestWithParam<std::function<TestDriverBase<Fptype>*()>> {
+/**
+ * Test class that's defining all tests to run with a Madgraph workflow.
+ * The tests are defined below using TEST_P.
+ * Instantiate them using:
+ * ```
+ * INSTANTIATE_TEST_SUITE_P( TestName,
+ *                           MadgraphTest,
+ *                           testing::Values( new TestImplementation, new TestImplementation2, ... ) );
+ * ```
+ */
+class MadgraphTest : public testing::TestWithParam<TestDriverBase*> {
 protected:
-  std::unique_ptr<TestDriverBase<Fptype>> testDriver;
-  using TestWithParamFptype = testing::TestWithParam<std::function<TestDriverBase<Fptype>*()>>;
-public:
-  MadgraphTestFptype() : TestWithParamFptype(), testDriver{ TestWithParamFptype::GetParam()() } {}
-  void madgraphTestBody();
+  std::unique_ptr<TestDriverBase> testDriver;
+
+  MadgraphTest() :
+    TestWithParam(),
+    testDriver{ GetParam() }
+  { }
 };
 
-typedef MadgraphTestFptype<float> MadgraphTestFloat;
 
-typedef MadgraphTestFptype<double> MadgraphTestDouble;
+// Since we link both the CPU-only and GPU tests into the same executable, we prevent
+// a multiply defined symbol by only compiling this in the non-CUDA phase:
+#ifndef __CUDACC__
+
+/// Compare momenta and matrix elements.
+/// This uses an implementation of TestDriverBase to run a madgraph workflow,
+/// and compares momenta and matrix elements with a reference file.
+TEST_P(MadgraphTest, CompareMomentaAndME)
+{
+  // Set to true to dump events:
+  constexpr bool dumpEvents = false;
+  constexpr double energy = 1500; // historical default, Ecms = 1500 GeV = 1.5 TeV (above the Z peak)
+  const double toleranceMomenta = std::is_same<double, fptype>::value ? 8.E-11 : 3.E-2;
+  const double toleranceMEs     = std::is_same<double, fptype>::value ? 1.E-6  : 2.E-3;
+
+  std::string dumpFileName = std::string("dump_")
+      + testing::UnitTest::GetInstance()->current_test_info()->test_suite_name()
+      + '.'
+      + testing::UnitTest::GetInstance()->current_test_info()->name()
+      + ".txt";
+  while (dumpFileName.find('/') != std::string::npos) {
+    dumpFileName.replace(dumpFileName.find('/'), 1, "_");
+  }
+  const std::string refFileName = testDriver->getRefFileName();
+
+  std::ofstream dumpFile;
+  if ( dumpEvents )
+  {
+    dumpFile.open(dumpFileName, std::ios::trunc);
+  }
+
+  // Read reference data
+  std::map<unsigned int, ReferenceData> referenceData;
+  if ( ! dumpEvents ) {
+    referenceData = readReferenceData(refFileName);
+  }
+  ASSERT_FALSE( HasFailure() ); // It doesn't make any sense to continue if we couldn't read the reference file.
+
+  // **************************************
+  // *** START MAIN LOOP ON #ITERATIONS ***
+  // **************************************
+  for (unsigned int iiter = 0; iiter < testDriver->niter; ++iiter)
+  {
+    testDriver->prepareRandomNumbers(iiter);
+
+    testDriver->prepareMomenta(energy);
+
+    testDriver->runSigmaKin(iiter);
+
+    // --- Run checks on all events produced in this iteration
+    for (std::size_t ievt = 0; ievt < testDriver->nevt && ! HasFailure(); ++ievt)
+    {
+      if (dumpEvents) {
+        ASSERT_TRUE(dumpFile.is_open()) << dumpFileName;
+        dumpFile << "Event " << std::setw(8) << ievt << "  "
+                 << "Batch " << std::setw(4) << iiter << "\n";
+        testDriver->dumpParticles(dumpFile, ievt, testDriver->nparticle, 15, ReferenceData());
+        // Dump matrix element
+        dumpFile << std::setw(4) << "ME" << std::scientific << std::setw(15+8)
+            << testDriver->getMatrixElement(ievt) << "\n" << std::endl << std::defaultfloat;
+        continue;
+      }
+
+      // Check that we have the required reference data
+      ASSERT_GT(referenceData.size(), iiter)
+        << "Don't have enough reference data for iteration " << iiter << ". Ref file:" << refFileName;
+      ASSERT_GT(referenceData[iiter].MEs.size(), ievt)
+        << "Don't have enough reference MEs for iteration " << iiter << " event " << ievt << ".\nRef file: " << refFileName;
+      ASSERT_GT(referenceData[iiter].momenta.size(), ievt)
+        << "Don't have enough reference momenta for iteration " << iiter << " event " << ievt << ".\nRef file: " << refFileName;
+      ASSERT_GE(referenceData[iiter].momenta[ievt].size(), testDriver->nparticle)
+        << "Don't have enough reference particles for iteration " << iiter << " event " << ievt << ".\nRef file: " << refFileName;
+
+      // This trace will help to understand the event that is being checked.
+      // It will only be printed in case of failures:
+      std::stringstream eventTrace;
+      eventTrace << "In comparing event " << ievt << " from iteration " << iiter << "\n";
+      testDriver->dumpParticles(eventTrace, ievt, testDriver->nparticle, 15, referenceData[iiter]);
+      eventTrace << std::setw(4) << "ME"   << std::scientific << std::setw(15+8)
+                 << testDriver->getMatrixElement(ievt) << "\n"
+                 << std::setw(4) << "r.ME" << std::scientific << std::setw(15+8)
+                 << referenceData[iiter].MEs[ievt] << std::endl << std::defaultfloat;
+      SCOPED_TRACE(eventTrace.str());
+
+      // Compare Momenta
+      for (unsigned int ipar = 0; ipar < testDriver->nparticle; ++ipar) {
+        std::stringstream momentumErrors;
+        for (unsigned int icomp = 0; icomp < mgOnGpu::np4; ++icomp) {
+          const double pMadg = testDriver->getMomentum(ievt, ipar, icomp);
+          const double pOrig = referenceData[iiter].momenta[ievt][ipar][icomp];
+          const double relDelta = fabs( (pMadg - pOrig)/pOrig );
+          if (relDelta > toleranceMomenta) {
+            momentumErrors << std::setprecision(15) << std::scientific << "\nparticle " << ipar << "\tcomponent " << icomp
+                << "\n\t madGraph:  " << std::setw(22) << pMadg
+                << "\n\t reference: " << std::setw(22) << pOrig
+                << "\n\t rel delta: " << std::setw(22) << relDelta << " exceeds tolerance of " << toleranceMomenta;
+          }
+        }
+        ASSERT_TRUE(momentumErrors.str().empty()) << momentumErrors.str();
+      }
+
+      // Compare ME:
+      EXPECT_NEAR(testDriver->getMatrixElement(ievt),
+          referenceData[iiter].MEs[ievt],
+          toleranceMEs * referenceData[iiter].MEs[ievt]);
+    }
+  }
+
+  if ( dumpEvents ) {
+    std::cout << "Event dump written to " << dumpFileName << std::endl;
+  }
+}
+
+#endif // __CUDACC__
 
 #endif /* MADGRAPHTEST_H_ */
